requirements.txt
streamlit
python-dotenv
pandas
docx
PyMuPDF
pytesseract
Pillow
requests
matplotlib
seaborn
google-generativeai
